Move source files to src/sequenceParser
* Disable the build of boost tests.
