+++ src/sequenceParser/Sequence.cpp
