+++ src/sequenceParser/common.hpp
