+++ src/sequenceParser/detail/FileStrings.cpp
