+++ src/sequenceParser/detector.cpp
