+++ src/sequenceParser/detector.hpp
