+++ src/sequenceParser/filesystem.cpp
@@ -1,6 +1,5 @@
-#include "common.hpp"
 #include "filesystem.hpp"
-#include "Sequence.hpp"
+
 #include "utils.hpp"
 
 #include "detail/analyze.hpp"
